requirements.txt
# HealthPay Reconciliation Agent - MCP Server
# Python 3.11+

# MCP SDK
mcp>=1.9.0

# Web Framework (Prompt Opinion HTTP transport)
fastapi>=0.115.0
uvicorn>=0.32.0

# FHIR Resources
fhir.resources>=7.0.0

# HTTP Client
httpx>=0.27.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# JWT handling (SHARP context)
PyJWT>=2.10.0

# Date handling
python-dateutil>=2.8.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# 0G Decentralized Compute
python-0g>=0.6.1
google-genai
